+++ lib/providers/settings/settings_provider.dart
@@ -8,15 +8,22 @@ final namespaceProvider = Provider<String>((ref) {
   return kReleaseMode ? 'release' : 'debug';
 });
 
-final settingsFutureProvider = FutureProvider<Settings>((ref) async {
+final fileSettingsFutureProvider = FutureProvider<Settings>((ref) async {
   final namespace = ref.watch(namespaceProvider);
   final sharedPrefs = await ref.read(sharedPreferencesProvider.future);
 
   return Settings(namespace, sharedPrefs);
 });
 
-final settingsProvider = ChangeNotifierProvider<Settings>((ref) {
-  return ref.watch(settingsFutureProvider).maybeWhen(
+final interfaceSettingsFutureProvider = FutureProvider<Settings>((ref) async {
+  final namespace = ref.watch(namespaceProvider);
+  final sharedPrefs = await ref.read(sharedPreferencesProvider.future);
+
+  return Settings(namespace, sharedPrefs);
+});
+
+final fileSettingsProvider = ChangeNotifierProvider<Settings>((ref) {
+  return ref.watch(fileSettingsFutureProvider).maybeWhen(
     data: (settings) {
       return settings;
     },
@@ -24,16 +31,28 @@ final settingsProvider = ChangeNotifierProvider<Settings>((ref) {
   );
 });
 
+final interfaceSettingsProvider = ChangeNotifierProvider<Settings>((ref) {
+  return ref.watch(interfaceSettingsFutureProvider).maybeWhen(
+      data: (settings) {
+        return settings;
+      },
+      orElse: () => throw Exception('Settings uninitialized.')
+  );
+});
+
 class Settings extends ChangeNotifier {
   Settings(this.namespace, this.sharedPreferences);
-
 
   final String namespace;
 
   final SharedPreferences sharedPreferences;
 
-  void setString(String key, String value) {
+  void setStringSilently(String key, String value) {
     sharedPreferences.setString(_key(key), value);
+  }
+
+  void setString(String key, String value) {
+    setStringSilently(key, value);
     notifyListeners();
   }